Apply `npm pkg fix`
I ran `npm pkg fix` in the repository in a clean state and these are the
changes it produced.

This follows the addition of this command in npm v9 as well as warnings
produced by the `npm publish` command which look like:

    $ npm publish --dry-run
    npm WARN publish npm auto-corrected some errors in your package.json
     when publishing.  Please run "npm pkg fix" to address these errors.
    npm WARN publish errors corrected:
    npm WARN publish "bin[jsfuzz]" script name was cleaned
    npm WARN publish "repository.url" was normalized to "[...]"
    [...]

Applying these changes directly to the version controlled manifest
avoids unexpected results from the auto-correction algorithm.